Created Raycasters folder and separated different classes in separate files
