+++ Source/SVONavigation/Public/Raycasters/SVORaycaster.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "SVONavigationTypes.h"
+
+#include <CoreMinimal.h>
+
+#include "SVORaycaster.generated.h"
+
+UCLASS( Abstract, NotBlueprintable, EditInlineNew )
+class SVONAVIGATION_API USVORaycaster : public UObject
+{
+    GENERATED_BODY()
+
+public:
+
+    bool HasLineOfSight( UObject * world_context, const FVector & from, const FVector & to, const FNavAgentProperties & nav_agent_properties ) const;
+    virtual bool HasLineOfSight( UObject * world_context, const FSVOVolumeNavigationData & volume_navigation_data, const FSVONodeAddress from, const FSVONodeAddress to, const FNavAgentProperties & nav_agent_properties ) const;
+    virtual bool HasLineOfSight( UObject * world_context, const FSVOVolumeNavigationData & volume_navigation_data, const FVector & from, const FVector & to, const FNavAgentProperties & nav_agent_properties ) const;
+};
